src/requirements.txt
# ─────────────────────────────────────────────────────
# OpinionFlow - Python Dependencies
# ─────────────────────────────────────────────────────
# FastAPI web framework with ASGI server
# ─────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ─────────────────────────────────────────────────────
# Database (SQLAlchemy + PostgreSQL)
# ─────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary

# ─────────────────────────────────────────────────────
# Authentication & Security
# ─────────────────────────────────────────────────────
argon2-cffi==23.1.0
authx>=1.1.0
pyjwt==2.10.1

# ─────────────────────────────────────────────────────
# Dependency Injection
# ─────────────────────────────────────────────────────
dishka==1.4.2

# ─────────────────────────────────────────────────────
# Data Validation & Settings
# ─────────────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ─────────────────────────────────────────────────────
# Message Broker (Kafka)
# ─────────────────────────────────────────────────────
faststream[kafka]

# ─────────────────────────────────────────────────────
# Logging
# ─────────────────────────────────────────────────────
loguru==0.7.3

# ─────────────────────────────────────────────────────
# Testing (optional, for development)
# ─────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# ─────────────────────────────────────────────────────
# Development Tools (optional)
# ─────────────────────────────────────────────────────
mypy==1.13.0
ruff==0.8.4
black==24.10.0
